refactor: :recycle: move validation to inputValidator file, auth middleware to middleware file
