+++ backend/routes/auth.js
@@ -1,12 +1,13 @@
 import express from 'express';
+
 import authController from '../controllers/authController.js';
-import authMiddleware from '../utilities/authMiddleware.js';
 import middleware from '../utilities/middleware.js';
+import inputValidator from '../utilities/inputValidator.js';
 
 const authRoute = express.Router();
 
-authRoute.post('/login', middleware.validateLogin, authController.login);
+authRoute.post('/login', inputValidator.validateLogin, authController.login);
 
-authRoute.get('/test', authMiddleware.authRequired, authController.test);
+authRoute.get('/test', middleware.authRequired, authController.test);
 
 export default authRoute;
